change monitor and actions generic
