+++ lib/schemas/common/actionCallbacks.js
@@ -1,7 +1,7 @@
 'use strict';
 
 // Edit page section topComponents callbacks
-const sectionCallbacks = ['reloadData', 'reloadSectionData'];
+const sectionCallbacks = ['reloadSectionData'];
 // Browse page topComponents callbacks
 const pageBrowseCallbacks = ['reloadBrowse'];
 
@@ -19,6 +19,7 @@ module.exports = ({ isBrowsePage, customCallbacks }) => ({
 			enum: [
 				'refresh',
 				'openLink',
+				'reloadData',
 				...customCallbacks || (isBrowsePage ? pageBrowseCallbacks : sectionCallbacks)
 			]
 		},
