+++ lib/schemas/common/generic-actions/action.js
@@ -11,10 +11,11 @@ const commonAttrs = {
 module.exports = {
 	type: 'object',
 	properties: {
+		componentAttributes: commonAttrs,
 		title: { $ref: 'schemaDefinitions#/definitions/stringPrefix' },
 		name: { type: 'string' },
 		type: { type: 'string' },
-		componentAttributes: commonAttrs,
+		kind: { type: 'string' },
 		sections: {
 			type: 'array',
 			items: { type: 'string' },
