+++ lib/schemas/common/generic-actions/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+const actions = require('./generic-actions');
+
+module.exports = actions;
