+++ lib/schemas/monitor/cards/BaseCard.js
@@ -2,7 +2,7 @@
 
 const { baseCard } = require('./cardNames');
 const fieldsGroup = require('../../../schemas/common-sections/sections/components/fieldGroup');
-const actions = require('./actions');
+const makeGenericActions = require('../../common/generic-actions');
 
 module.exports = {
 	if: {
@@ -13,7 +13,7 @@ module.exports = {
 	then: {
 		properties: {
 			component: { type: 'string', const: baseCard },
-			actions,
+			actions: makeGenericActions({ customCallbacks: [] }),
 			fieldsGroup
 		},
 		required: ['component', 'fieldsGroup'],
