--- lib/schemas/preview/actions/index.js
@@ -1,5 +1,0 @@
-'use strict';
-
-const actions = require('./actions');
-
-module.exports = actions;
