+++ lib/schemas/preview/schema.js
@@ -5,8 +5,9 @@ const themes = require('../common/themes');
 const sectionNames = require('./sectionsNames');
 const makeSections = require('../common-sections');
 const header = require('../edit-new/modules/headerProperties');
-const actions = require('./actions');
+const makeGenericActions = require('../common/generic-actions');
 
+const customCallbacks = ['removeRow', 'reloadRow', 'reloadBrowse'];
 
 module.exports = {
 	type: 'object',
@@ -14,7 +15,7 @@ module.exports = {
 		...properties,
 		header,
 		themes,
-		actions,
+		actions: makeGenericActions({ customCallbacks }),
 		root: { const: 'Preview' },
 		source: { $ref: 'schemaDefinitions#/definitions/endpoint' },
 		sourceEndpointParameters: { $ref: 'schemaDefinitions#/definitions/endpointParameters' },
