+++ tests/mocks/schemas/monitor.yml
@@ -122,45 +122,51 @@ fields:
     translateLabel: false
     actions:
       - name: testEndpoint
+        type: endpoint
         callback: refresh
-        endpoint:
-          service: serviceName
-          namespace: namespaceName
-          method: methodName
-          resolve: false
-        endpointParameters:
-          - name: id
-            target: path
-            value:
-              dynamic: id
+        componentAttributes:
+          icon: box
+          endpoint:
+            service: serviceName
+            namespace: namespaceName
+            method: methodName
+            resolve: false
+          endpointParameters:
+            - name: id
+              target: path
+              value:
+                dynamic: id
 
       - name: testLink
-        path: /test/{id}
-        target: _self
+        type: link
         conditions:
           showWhen:
             - - name: isEqualTo
                 field: name
                 referenceValue: test
-        endpointParameters:
-          - name: id
-            target: path
-            value:
-              dynamic: id
+        componentAttributes:
+          path: /test/{id}
+          linkTarget: _self
+          endpointParameters:
+            - name: id
+              target: path
+              value:
+                dynamic: id
 
       - name: testForm
-        endpoint:
-          service: serviceName
-          namespace: namespaceName
-          method: methodName
-          resolve: false
+        type: form
         callback: reloadData
-        schema:
-          - name: fieldNameOne
-            component: Input
-
-          - name: fieldNameTwo
-            component: Switch
+        componentAttributes:
+          endpoint:
+            service: serviceName
+            namespace: namespaceName
+            method: methodName
+            resolve: false
+          fields:
+            - name: fieldNameOne
+              component: Input
+            - name: fieldNameTwo
+              component: Switch
 
     fieldsGroup:
       - name: info
